Make all the LFCLK ticks macros convert first to 64 bit and then to 32 bit to prevent pre-mature overflow

+++ codebase/util/nrf_util.h
@@ -100,13 +100,13 @@ typedef enum
 #define LFCLK_FREQ      32768
 
 /** Macro to find out the number of LFCLK ticks for the passed time in milli-seconds */
-#define LFCLK_TICKS_MS(ms)                ((uint32_t) ROUNDED_DIV( (LFCLK_FREQ*ms) , 1000) )
+#define LFCLK_TICKS_MS(ms)                ((uint32_t) ROUNDED_DIV( (LFCLK_FREQ*(uint64_t)ms) , 1000) )
 /** Macro to find out the number of LFCLK ticks for the passed time in multiples of 0.625 ms */
-#define LFCLK_TICKS_625(ms)               ((uint32_t) ROUNDED_DIV( (LFCLK_FREQ*ms) , 1600) )
+#define LFCLK_TICKS_625(ms)               ((uint32_t) ROUNDED_DIV( (LFCLK_FREQ*(uint64_t)ms) , 1600) )
 /** Macro to find out the number of LFCLK ticks for the passed time in multiples of 0.977 ms */
 #define LFCLK_TICKS_977(ms)               ((uint32_t) (32*ms) )
 /** Macro to find out the number of LFCLK ticks for the passed time in multiples of 1.25 ms */
-#define LFCLK_TICKS_1250(ms)              ((uint32_t) ROUNDED_DIV( (LFCLK_FREQ*ms) , 800) )
+#define LFCLK_TICKS_1250(ms)              ((uint32_t) ROUNDED_DIV( (LFCLK_FREQ*(uint64_t)ms) , 800) )
 
 void nrf_util_critical_region_enter(uint8_t * is_critical_entered);
 void nrf_util_critical_region_exit(uint8_t critical_entered);
